instance-segmentation/Mask_RCNN/requirements.txt
matplotlib==3.10.7
numpy==2.3.3
pillow==11.3.0
pycocotools==2.0.10
torch==2.5.1+cu121
torchvision==0.20.1+cu121